# docs/openapi/assinatura.yaml
openapi: 3.0.3
info:
  title: Assinatura Service API
  description: |
    Endpoints de solicitacao e consulta de assinatura. A assinatura nasce
    AGUARDANDO_PAGAMENTO, torna-se ATIVA na aprovacao do pagamento e daí segue o
    ciclo de renovacao automatica (EM_RENOVACAO, SUSPENSA, CANCELADA).

    Todas as rotas exigem JWT emitido por `POST /auth/login` (ver ADR 0003). O
    dono da assinatura vem sempre do token: o corpo da solicitacao carrega
    apenas o plano, e a consulta so devolve a assinatura ao proprio dono ou a um
    administrador.

    O cancelamento tem contrato proprio:
    `docs/openapi/cancelamento-assinatura.yaml`.
  version: 0.3.0
servers:
  - url: http://localhost:18080
    description: Local (docker-compose)
tags:
  - name: Assinatura
    description: Solicitacao e consulta de assinatura

security:
  - bearerAuth: []

paths:
  /assinaturas:
    post:
      tags: [Assinatura]
      summary: Solicita uma assinatura
      description: |
        Cria uma assinatura em AGUARDANDO_PAGAMENTO para o usuario do token.
        Regra "um usuario, uma assinatura aberta": se ja existe assinatura em
        AGUARDANDO_PAGAMENTO ou ATIVA para esse usuario, retorna 409.

        O dono nao vem no corpo: campos extras enviados sao ignorados. A
        assinatura nasce sempre AGUARDANDO_PAGAMENTO, e dataInicio e
        dataExpiracao sao definidas na ativacao, quando o pagamento e aprovado.
      operationId: solicitarAssinatura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssinaturaSolicitacao'
            examples:
              premium:
                summary: Solicitacao de plano PREMIUM
                value:
                  plano: PREMIUM
      responses:
        '202':
          description: Assinatura criada em AGUARDANDO_PAGAMENTO.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssinaturaCriada'
              example:
                id: "770e8400-e29b-41d4-a716-446655440000"
                status: AGUARDANDO_PAGAMENTO
        '400':
          description: Corpo invalido (campos ausentes, plano nao suportado, uuid malformado).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                status: 400
                title: Requisicao invalida
                errors:
                  - campo: plano
                    motivo: deve ser um dos planos suportados (BASICO, PREMIUM, FAMILIA)
        '401':
          description: Token ausente, expirado ou invalido. Corpo vazio.
        '403':
          description: |
            Token valido sem usuario de dominio ligado (por exemplo, o
            administrador do seed), que portanto nao pode assinar. Corpo vazio.
        '404':
          description: |
            O usuario do token nao corresponde a nenhum usuario cadastrado.
            Corpo vazio.
        '409':
          description: Usuario ja possui assinatura aberta. Corpo vazio.

  /assinaturas/{uuid}:
    get:
      tags: [Assinatura]
      summary: Consulta uma assinatura pelo uuid publico
      description: |
        Devolve a representacao completa da assinatura ao proprio dono.
        Assinatura de outro usuario retorna 403; ROLE_ADMIN consulta qualquer
        uma. dataInicio e dataExpiracao sao nulas enquanto a assinatura aguarda
        pagamento.
      operationId: consultarAssinatura
      parameters:
        - name: uuid
          in: path
          required: true
          description: uuid publico da assinatura.
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Assinatura encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assinatura'
              examples:
                aguardando:
                  summary: Aguardando pagamento (datas nulas)
                  value:
                    id: "770e8400-e29b-41d4-a716-446655440000"
                    usuarioId: "550e8400-e29b-41d4-a716-446655440000"
                    plano: PREMIUM
                    dataInicio: null
                    dataExpiracao: null
                    status: AGUARDANDO_PAGAMENTO
                    renovacaoAutomatica: true
                    acessoAte: null
                ativa:
                  summary: Assinatura ativa, renovando ao fim do ciclo
                  value:
                    id: "770e8400-e29b-41d4-a716-446655440000"
                    usuarioId: "550e8400-e29b-41d4-a716-446655440000"
                    plano: PREMIUM
                    dataInicio: "2025-03-10"
                    dataExpiracao: "2025-04-10"
                    status: ATIVA
                    renovacaoAutomatica: true
                    acessoAte: "2025-04-10"
                cancelamentoAgendado:
                  summary: Cancelada pelo cliente, acesso ate o fim do ciclo
                  value:
                    id: "770e8400-e29b-41d4-a716-446655440000"
                    usuarioId: "550e8400-e29b-41d4-a716-446655440000"
                    plano: PREMIUM
                    dataInicio: "2025-03-10"
                    dataExpiracao: "2025-04-10"
                    status: ATIVA
                    renovacaoAutomatica: false
                    acessoAte: "2025-04-10"
                cancelada:
                  summary: Cancelamento efetivado no vencimento
                  value:
                    id: "770e8400-e29b-41d4-a716-446655440000"
                    usuarioId: "550e8400-e29b-41d4-a716-446655440000"
                    plano: PREMIUM
                    dataInicio: "2025-03-10"
                    dataExpiracao: "2025-04-10"
                    status: CANCELADA
                    renovacaoAutomatica: false
                    acessoAte: "2025-04-10"
        '401':
          description: Token ausente, expirado ou invalido. Corpo vazio.
        '403':
          description: Assinatura pertence a outro usuario. Corpo vazio.
        '404':
          description: Assinatura nao encontrada. Corpo vazio.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `POST /auth/login` (ver `docs/openapi/auth.yaml`). O
        subject e o email do cliente e a claim `usuarioId` identifica o usuario
        de dominio dono das assinaturas.
  schemas:
    Plano:
      type: string
      description: |
        Plano da assinatura. Valor mensal (BigDecimal), mapeado internamente e
        exposto nos eventos de assinatura e renovacao:
          - BASICO: R$ 19,90 (19.90)
          - PREMIUM: R$ 39,90 (39.90)
          - FAMILIA: R$ 59,90 (59.90)
        Validado contra os planos suportados.
      enum:
        - BASICO
        - PREMIUM
        - FAMILIA
      example: PREMIUM
    StatusAssinatura:
      type: string
      description: |
        Status do ciclo de vida da assinatura:
          - AGUARDANDO_PAGAMENTO: recem-solicitada, sem vigencia definida.
          - ATIVA: pagamento aprovado, acesso liberado ate dataExpiracao.
          - PAGAMENTO_RECUSADO: primeira cobranca recusada.
          - EM_RENOVACAO: ciclo vencido, cobranca da renovacao em andamento.
            O acesso permanece liberado enquanto a cobranca e decidida.
          - SUSPENSA: acesso interrompido apos as tres tentativas de cobranca
            da renovacao se esgotarem.
          - CANCELADA: ciclo de vida encerrado, por cancelamento imediato ou
            pela efetivacao do opt-out no vencimento.
        A solicitacao (POST) produz apenas AGUARDANDO_PAGAMENTO.
      enum:
        - AGUARDANDO_PAGAMENTO
        - ATIVA
        - PAGAMENTO_RECUSADO
        - EM_RENOVACAO
        - SUSPENSA
        - CANCELADA
      example: AGUARDANDO_PAGAMENTO
    AssinaturaSolicitacao:
      type: object
      description: |
        Input da solicitacao de assinatura. O plano e a unica entrada; o dono
        vem da claim `usuarioId` do token, e id, status, dataInicio e
        dataExpiracao sao derivados pelo servidor.
      required: [plano]
      properties:
        plano:
          $ref: '#/components/schemas/Plano'
    AssinaturaCriada:
      type: object
      description: Resposta da solicitacao (202).
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
          description: uuid publico da assinatura criada.
        status:
          $ref: '#/components/schemas/StatusAssinatura'
    Assinatura:
      type: object
      description: Representacao completa da assinatura (GET).
      required: [id, usuarioId, plano, status, renovacaoAutomatica]
      properties:
        id:
          type: string
          format: uuid
          description: uuid publico da assinatura.
        usuarioId:
          type: string
          format: uuid
          description: uuid publico do usuario dono da assinatura.
        plano:
          $ref: '#/components/schemas/Plano'
        dataInicio:
          type: string
          format: date
          nullable: true
          description: Definida na ativacao. Nula enquanto aguarda pagamento.
        dataExpiracao:
          type: string
          format: date
          nullable: true
          description: Definida na ativacao. Nula enquanto aguarda pagamento.
        status:
          $ref: '#/components/schemas/StatusAssinatura'
        renovacaoAutomatica:
          type: boolean
          description: |
            `true` enquanto a assinatura renova ao fim de cada ciclo. Vira
            `false` no cancelamento e nunca volta a `true`.
        acessoAte:
          type: string
          format: date
          nullable: true
          description: |
            Fim do ciclo corrente (`fimCiclo`), ate quando o servico permanece
            disponivel. Nulo antes da primeira ativacao. Com status ATIVA e
            `renovacaoAutomatica: false`, e a data em que o cancelamento e
            efetivado pela varredura de vencimentos.
    Problem:
      type: object
      description: |
        Detalhe de erro (RFC 7807/9457). No 400, o array `errors` lista os campos
        invalidos para o cliente corrigir a requisicao. 404 e 409 retornam corpo
        vazio.
      properties:
        status:
          type: integer
          description: Codigo de status HTTP.
        title:
          type: string
          description: Titulo curto do erro.
        errors:
          type: array
          description: Campos invalidos (presente no 400).
          items:
            type: object
            required: [campo, motivo]
            properties:
              campo:
                type: string
                description: Nome do campo invalido.
              motivo:
                type: string
                description: Motivo da invalidade.